+++ src/components/native-select/native-select.tsx
@@ -1,21 +1,25 @@
-import { Component, h, ComponentInterface, State, Prop, Host, Element } from '@stencil/core';
+import { Component, h, ComponentInterface, State, Prop, Host, Element, Method } from '@stencil/core';
 import { SelectArrow } from '../select/select-arrow';
 import { WcsSize } from "../../shared-types";
 
 export type WcsNativeSelectSize = Extract<WcsSize, 'l' | 'm'>; // todo: move into common file with custom select
 
-
-
 /**
- * The `wcs-native-select` component is designed to accept a native <select> element as a slotted child. This choice
- * allows developers to bind the <select> element using the framework of their choice, without the need to re-expose all the
- * properties of the <select> and <option> elements in this component.
+ * The `wcs-native-select` component is designed to accept a native `<select>` element as a slotted child. This choice
+ * allows developers to bind the `<select>` element using the framework of their choice, without the need to re-expose all the
+ * properties of the `<select>` and `<option>` elements in this component.
  *
- * The component wraps the native <select> element and provides custom styles and behavior, while preserving the native
+ * The component wraps the native `<select>` element and provides custom styles and behavior, while preserving the native
  * functionality and accessibility.
  *
- * Example usage:
+ * ### ✅ Guidance
  *
+ * - To have a placeholder, you must have an option as child which has `selected` attribute and `disabled`
+ * attribute. You can add the `hidden` attribute to don't show the placeholder option in the options overlay.
+ *
+ * ### Example usage
+ *
+ * ```html
  * <wcs-native-select>
  *   <select>
  *     <option value="option1">Option 1</option>
@@ -23,6 +27,12 @@ export type WcsNativeSelectSize = Extract<WcsSize, 'l' | 'm'>; // todo: move int
  *     <option value="option3">Option 3</option>
  *   </select>
  * </wcs-native-select>
+ * ```
+ *
+ * ### Note
+ * - We did not find a way to detect when the select is reset, if you want to apply the placeholder style when the
+ * select is reset, you have to call the `updateStyles()` method manually.
+ * - It is strongly recommended to use select-native when you don't have to support the multi-selection feature
  */
 @Component({
     tag: 'wcs-native-select',
@@ -38,7 +48,7 @@ export class NativeSelect implements ComponentInterface {
      *
      * The default value is 'm'.
      */
-    @Prop({ reflect: true }) size: WcsNativeSelectSize = 'm';
+    @Prop({reflect: true}) size: WcsNativeSelectSize = 'm';
 
     @Element() private el!: HTMLWcsNativeSelectElement;
 
@@ -69,12 +79,23 @@ export class NativeSelect implements ComponentInterface {
             if (hasSpiedAttrMutation) {
                 this.updateHostAttributeWithSlottedSelect();
             }
+
+            // A workaround to detect when the select is reset by using a class change on the select element. For example
+            // angular add/remove ng-pristine and ng-dirty classes on the form controls when the form is reset.
+            // Not the best solution but it works since we cannot detect the reset event on the native select element.
+            if (mutations.filter(m => m.attributeName === 'class').length > 0) {
+                this._updateStyles();
+            }
         });
         this.updateHostAttributeWithSlottedSelect();
         this.observer.observe(this.selectElement, {attributes: true});
     }
 
     private onSelectedOptionChange(): void {
+        this._updateStyles();
+    }
+
+    private _updateStyles() {
         if (this.isPlaceholderOptionSelected()) {
             this.applyPlaceholderStylesOnNativeSlottedSelectElement();
         } else {
@@ -129,6 +150,18 @@ export class NativeSelect implements ComponentInterface {
         this.selectElement?.removeEventListener("change", () => {
             this.onSelectedOptionChange();
         });
+    }
+
+    /**
+     * Use this method to force the component to update its styles. It can be useful when the select is reset (with a placeholder).
+     */
+    @Method()
+    async updateStyles() {
+        // Note : we try to automatically detect if the select is reset with the onreset event and with a mutation
+        // observer on the select element. But it seems that the onreset event is not fired when the select is reset.
+        // and the mutation observer is not fired when the select is reset with the reset() method on the form.
+        // So we have to let the user call this method manually when he reset the select.
+        this._updateStyles();
     }
 
     render() {